+++ src/embed/command/runSql.ts
@@ -21,31 +21,31 @@ export async function runSelectedSql(
     (process, token) => {
       return new Promise<void>(async (resolve) => {
         process.report({ message: "Preparing to execute sql", increment: 0 });
-        let [results, fields] = await execSelectAsyncProcess(
-          config,
-          schemaName,
-          sql,
-          process,
-          1,
-          100
-        );
-        // get results type then goto result handler strategy
-        let constructorName;
-        if (results instanceof Array) {
-          constructorName = Object.getPrototypeOf(results[0]).constructor.name;
-        } else {
-          constructorName = Object.getPrototypeOf(results).constructor.name;
-        }
-        let handler = resultHandlerStrategy.get(constructorName);
-        if (handler) {
-          handler({ results: results, fields: fields, sql: sql });
-        } else {
-          Logger.attension(
-            String.raw`There's no result handler for this sql, Please submit to issue
-        url: https://github.com/AethLi/elegance_mysql/issues`
-          );
-        }
-        resolve();
+        execSelectAsyncProcess(config, schemaName, sql, process, 1, 100)
+          .then(([results, fields]) => {
+            // get results type then goto result handler strategy
+            let constructorName;
+            if (results instanceof Array) {
+              constructorName = Object.getPrototypeOf(results[0]).constructor
+                .name;
+            } else {
+              constructorName = Object.getPrototypeOf(results).constructor.name;
+            }
+            let handler = resultHandlerStrategy.get(constructorName);
+            if (handler) {
+              handler({ results: results, fields: fields, sql: sql });
+            } else {
+              Logger.attension(
+                String.raw`There's no result handler for this sql, Please submit to issue
+                    url: https://github.com/AethLi/elegance_mysql/issues`
+              );
+            }
+            resolve();
+          })
+          .catch((e) => {
+            Logger.error(e.message, e);
+            resolve();
+          });
       });
     }
   );