+++ src/capability/logService.ts
@@ -73,11 +73,11 @@ export class Logger {
   }
 
   public static error(message: string, error?: Error | string) {
+    window.showErrorMessage(message);
     if (this.logLevel === "NONE") {
       return;
     }
     this.logMessage(message, "ERROR");
-    window.showErrorMessage(message);
     if (typeof error === "string") {
       // Errors as a string usually only happen with
       // plugins that don't return the expected error.